+++ tests/unit/test_utils.py
@@ -7,9 +7,9 @@ import scipy.stats as ss
 import elfi
 from elfi.examples.ma2 import get_model
 from elfi.methods.bo.utils import minimize, stochastic_optimization
-from elfi.methods.utils import (DensityRatioEstimation, GMDistribution, normalize_weights,
-                                numgrad, numpy_to_python_type, sample_object_to_dict,
-                                weighted_sample_quantile, weighted_var)
+from elfi.methods.density_ratio_estimation import DensityRatioEstimation
+from elfi.methods.utils import (GMDistribution, normalize_weights, numgrad, numpy_to_python_type,
+                                sample_object_to_dict, weighted_sample_quantile, weighted_var)
 from elfi.model.extensions import ModelPrior
